+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Version 25.03.XX
 Fixes:
+- [core] Fix user analytics widget chart
 - [core] Fix mongo connection url parsing
 
 ## Version 25.03.10